reorganize helix-loader under helix-core
This commit is a rather large refactor that places helix-loader underneath
helix-core in the crate dependency chain. This includes moving some code
related to config (merge_toml_values), which is necessary because loader
may wish to read the languages.toml on its own when fetching or building
grammars. I tried to move as little as possible: the language configuration
remains in helix-core.