Return transactions from commands
This probably also needs to be able to return selection transformations
because commands will change the text via transactions as an
after-effect but the selections will be moved first and be
out-of-bounds on inserts past the end of the original doc.

Either that or I guess we could wrap up every change to selections
in transactions that have an empty set of changes?