(5176) Cross injection layers in tree-sitter motions
Tree-sitter motions didn't work well for languages with injections
such as Vue or JavaScript inside HTML since only the root layer's
syntax-tree was considered. This change uses the range information for
each injected language layer to find the appropriate syntax tree
so that tree-sitter motions (A-p, A-i, A-o, A-n) work within injected
content.