Tune default configuration options
This commit is specific to this driver branch: it's equivalent to
the following TOML configuration:

    [editor]
    idle-timeout = 250
    soft-wrap = true
    rainbow-brackets = true

Why change the defaults in the code and not just use a config file?
I end up running a lot of PRs through `cargo run` when reviewing and
old branches might not have these options.

The same could be accomplished by creating a wrapper shell script
around `hx` that calls `hx -c /path/to/custom-config.toml $@` but
it's more work to put that together than to just change these values
in the code.