Add "fallback commands"
These are meant for the likes of `mi` and `ma` special commands and
might be useful for `]`/`[` in the future. When no command is bound to
a key, we check if there is a fallback command and if so execute that
given the last entered character.

Currently this feels a little hacky. I'm duplicating some code in the
keymap file and the FallbackCommand is very similar to MappableCommands,
just a different signature. I wonder if I should just add fallbacks
into MappableCommand's enum.